chore: remove dead constants, narrow widened snapshot types, name frozen selectors, raise stream frame floor (#8)
* chore(hub): delete 58 unreferenced enum-style string constants

keys.go carried 65 SCREAMING_SNAKE string constants shadowing protobuf enums.
58 of them have no Go reference anywhere in the tree, tests included, and their
literal values appear in no json or testdata file either, so removing them
cannot change behaviour.

The 7 that survive are referenced: ServiceBondRole{Worker,Verifier} and the five
ModelSupportDeactivate* values that SupportDeactivateCursorState.reason compares
against.

Also drops MaxSupportedProfilesPerOperator. It had zero references while an
identically named governance parameter carries the live bound, which is the
shape that already bit us once: MinServiceBond was an unregistered duplicate of
its parameter and had silently drifted from it (500_000 against 1_000_000)
before being reconciled. The two values here happen to agree today, but nothing
links them, so changing the parameter would leave the constant behind.

MaxProfilesPerModel is deliberately kept - it has two live references.

Five const blocks are left empty by the removal and are deleted with their
headers.

* refactor(hub): narrow ProfileStateSnapshot tier and supporter count to uint32

ResourceTier and ActiveSupporterCount are uint32 on ProfileState. The snapshot
widened both to uint64 for no stated reason, so every producer inserted a
widening conversion and the one consumer had to guard and narrow back:

    if profile.ResourceTier > math.MaxUint32 { ... }        // source is uint32
    bucket, err := RewardBucketForResourceTier(uint32(profile.ResourceTier))

That guard is unreachable - the value originates as uint32 - and redundant:
RewardBucketForResourceTier already rejects anything outside the registered
1..5 range and returns the more precise error. Both are removed along with the
now-unused math import.

The snapshot is a plain Go struct passed across the module boundary, not a
persisted or wire type, so no encoding changes.

Neither field has a reader outside hub today; they are left in place rather than
deleted because the support aggregation they belong to is being reworked.

* refactor(shared): name the four frozen execution and verification selectors

runtime_class, judgment_function_version, canonical_encoding_version and
metric_aggregate_proof_version each accept exactly one value in V1. Those values
were inline literals at nine sites across three files, and the registration path
in hub carried its own copy of every string while the task admission and
verification paths carried a second:

    x/hub/types/model.go              registration accepts the profile
    x/task/keeper/task_admission.go   admission re-checks it
    x/task/keeper/verification_runtime.go

Two independently typed copies of a consensus string is the failure mode where a
typo in one lets a profile register and then be refused at admission, with
nothing to catch the divergence. The strings also feed the registration hash
preimage, so a silent drift would change a digest rather than just a comparison.

They move to x/shared/types, which both modules already import, as named
constants. Values are unchanged byte for byte - the golden digest vectors cover
this and still pass.

They stay strings rather than becoming enums: the same names appear in the
off-chain profile manifest that non-Go participants read, and switching to
numeric enums would change the hash preimage for no proportional gain, since
profile rows are capped per model.

* fix(task): raise min_output_stream_frame_bytes default to 256

Closes #6.

At 16 bytes the fixed cost of a signed output frame - MMR root 32, worker
signature 64, seq 8 - is roughly twelve times the payload it carries, and each
frame also costs one signing operation, one durable evidence write and one MMR
leaf. A 2915-byte output produced around 182 frames.

256 is the agreed value. It is a floor on non-final frames, so it also sets the
floor on time to first frame: at 30 token/s with CJK text a frame lands every
2-3 seconds, while the payload per frame already exceeds what a reader consumes
in that time. Going further - 1 KiB and up - pushes first-frame latency into the
tens of seconds, which costs more in delivery experience than it saves in
overhead. The final frame is exempt and carries whatever remains.

Both the code default and the localnet genesis seed move together so a freshly
seeded chain and a default-constructed params value agree.

The two TaskParamsHashV1 anchors in params_test.go move with it. That is the
intended behaviour of those tests, not incidental: they exist so that changing
any default is visible as a digest change. The preimage itself is unchanged -
TestTaskParamsHashCoversEveryField still walks all 84 field mutations without a
collision, and the standalone params fixture is unaffected because it frames its
own listed values rather than the code defaults.

+++ x/hub/keeper/hub_keeper.go
@@ -119,9 +119,9 @@ func (k Keeper) GetProfileState(ctx sdk.Context, modelID string, profileVersion 
 	}
 	return types.ProfileStateSnapshot{
 		ModelID: state.ModelId, ProfileVersion: state.ProfileVersion, Status: state.Status,
-		TaskTypes: append([]shared.TaskType(nil), state.TaskTypes...), ResourceTier: uint64(state.ResourceTier),
+		TaskTypes: append([]shared.TaskType(nil), state.TaskTypes...), ResourceTier: state.ResourceTier,
 		MinStake: state.MinStake, ChallengeOpenWindowBlocks: state.ChallengeOpenWindowBlocks,
-		ActiveSupporterCount: uint64(state.ActiveSupporterCount), ActiveSupportStake: state.ActiveSupportStake,
+		ActiveSupporterCount: state.ActiveSupporterCount, ActiveSupportStake: state.ActiveSupportStake,
 		StatusSource: state.StatusSource, ExecutionSnapshot: execution, ExecutionSnapshotHash: executionHash,
 		PricingProfile: state.PricingProfile, RefPrice: state.RefPrice,
 	}, true
